requirements.txt
playwright>=1.47
pydantic>=2.5
imapclient>=2.3
beautifulsoup4>=4.12
rich>=13.7
tenacity>=8.4
python-dotenv>=1.0
